modify attention

+++ stmt2vec_attention/optimization.py
@@ -5,7 +5,7 @@ import numpy as np
 
 def adagrad(parameter, parameter_gradient, learning_rate=.05, fudge_factor=1e-10, clip_threshold=1):
     clipped_gradient = T.clip(parameter_gradient, -clip_threshold, clip_threshold)
-    adagrad_historical = theano.shared(np.zeros(parameter.get_value().shape, dtype=np.float), "adagrad_historical")
+    adagrad_historical = theano.shared(np.zeros(parameter.get_value().shape, dtype=np.float32), "adagrad_historical")
     next_adagrad = adagrad_historical + T.pow(clipped_gradient, 2)
     adagrad_update = adagrad_historical, next_adagrad
     update = learning_rate / T.sqrt(fudge_factor + next_adagrad) * clipped_gradient
@@ -16,7 +16,7 @@ def adagrad(parameter, parameter_gradient, learning_rate=.05, fudge_factor=1e-10
 
 def rmsprop(parameter, parameter_gradient, learning_rate=.05, fudge_factor=1e-10, rho=.9, clip_threshold=1):
     clipped_gradient = T.clip(parameter_gradient, -clip_threshold, clip_threshold)
-    rmsprob_moving_avg = theano.shared(np.ones(parameter.get_value().shape, dtype=np.float) * 0, "rmsprop_historical")
+    rmsprob_moving_avg = theano.shared(np.ones(parameter.get_value().shape, dtype=np.float32) * 0, "rmsprop_historical")
     next_rmsprop_avg = rho * rmsprob_moving_avg + (1. - rho) * T.pow(clipped_gradient, 2)
     update = rmsprob_moving_avg, next_rmsprop_avg
     grad_step = learning_rate / T.sqrt(fudge_factor + next_rmsprop_avg) * clipped_gradient
@@ -60,8 +60,8 @@ def nesterov_rmsprop_multiple(parameters, parameter_gradients, learning_rate=.00
     return updates
 
 def nesterov_rmsprop(parameter, parameter_gradient, learning_rate, momentum, fudge_factor=1e-10, rho=.9):
-    memory = theano.shared(np.zeros_like(parameter.get_value(), dtype=np.float), name="nesterov_momentum")
-    rmsprop_moving_avg = theano.shared(np.zeros(parameter.get_value().shape, dtype=np.float), "rmsprop_historical")
+    memory = theano.shared(np.zeros_like(parameter.get_value(), dtype=np.float32), name="nesterov_momentum")
+    rmsprop_moving_avg = theano.shared(np.zeros(parameter.get_value().shape, dtype=np.float32), "rmsprop_historical")
 
     next_rmsprop_avg = rho * rmsprop_moving_avg + (1. - rho) * T.pow(parameter_gradient, 2)
     memory_update = memory, momentum * memory + learning_rate / T.sqrt(fudge_factor + next_rmsprop_avg) * parameter_gradient
